DatasetsList container - mapStateToProps moved to anonymous function in connect. Renamed Datasets.js to DatasetsList.js to match component name change
